resolved sonar fixes

+++ src/main/java/com/drajer/bsa/controller/NotificationContextController.java
@@ -8,8 +8,6 @@ import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.List;
 import java.util.Map;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -29,7 +27,6 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 public class NotificationContextController {
 
-  private final Logger logger = LoggerFactory.getLogger(NotificationContextController.class);
   private final NotificationContextService notificationContextService;
 
   /**
@@ -53,11 +50,11 @@ public class NotificationContextController {
             notificationContextData.getId(), notificationContextData.getFhirServerBaseUrl(),
             notificationContextData.getNotificationResourceId(),
                 notificationContextData.getPatientId());
-    if (notificationContextDetails != null) {
+    if (notificationContextDetails != null && !notificationContextDetails.isEmpty()) {
       for (NotificationContext notificationContext : notificationContextDetails) {
         notificationContextService.delete(notificationContext);
-        return "NotificationContext deleted successfully.";
       }
+      return "NotificationContext deleted successfully.";
     }
     response.sendError(HttpServletResponse.SC_NOT_FOUND, "NotificationContext Not found");
     return "NotificationContext Not found";
